Fixed updating with an installed version
Tested and now seems to work when updating from a previous version.

Also seems to work fine when installing fresh.

+++ craft.sh
@@ -46,7 +46,7 @@ check_version () {
    VERSION=${VERSION%|*}
    VERSION=${VERSION##*|}
 
-   if [ -f ".region" ]; then
+   if [ -f ".version" ]; then
      INSTALLED=$(cat .version)
    else
      INSTALLED="Not installed"
@@ -75,19 +75,24 @@ move_files_and_cleaup () {
  echo -e "${GREEN}Moving files & running cleanup ...${WHITE}"
  [ ! -d "$TARGET_PATH/Bin" ] && mkdir -p $TARGET_PATH/Bin
 
- echo -e "${GREEN}Extracting Unity files....${WHITE}\n"
- tar -xf tmp/Unity.tar.xz -C tmp Editor/Data/PlaybackEngines/LinuxStandaloneSupport/Variations/linux64_withgfx_nondevelopment_mono/LinuxPlayer Editor/Data/PlaybackEngines/LinuxStandaloneSupport/Variations/linux64_withgfx_nondevelopment_mono/Data/MonoBleedingEdge/
+ if [ -f "tmp/Unity.tar.xz" ]; then
+   echo -e "${GREEN}Extracting Unity files....${WHITE}\n"
+   tar -xf tmp/Unity.tar.xz -C tmp Editor/Data/PlaybackEngines/LinuxStandaloneSupport/Variations/linux64_withgfx_nondevelopment_mono/LinuxPlayer Editor/Data/PlaybackEngines/LinuxStandaloneSupport/Variations/linux64_withgfx_nondevelopment_mono/Data/MonoBleedingEdge/
 
- UNITY_PATH=tmp/Editor/Data/PlaybackEngines/LinuxStandaloneSupport/Variations/linux64_withgfx_nondevelopment_mono
+   UNITY_PATH=tmp/Editor/Data/PlaybackEngines/LinuxStandaloneSupport/Variations/linux64_withgfx_nondevelopment_mono
+
+   if [ ! -f "$TARGET_PATH/Bin/Hearthstone.x86_64" ]; then
+     cp $UNITY_PATH/LinuxPlayer $TARGET_PATH/Bin/Hearthstone.x86_64
+     cp -r $UNITY_PATH/Data/MonoBleedingEdge $TARGET_PATH/Bin/Hearthstone_Data
+   fi
+
+ fi
+
+ cp -r $TARGET_PATH/Hearthstone.app/Contents/Resources/Data $TARGET_PATH/Bin/Hearthstone_Data
+ cp -r $TARGET_PATH/Hearthstone.app/Contents/Resources/'unity default resources' $TARGET_PATH/Bin/Hearthstone_Data/Resources
+ cp -r $TARGET_PATH/Hearthstone.app/Contents/Resources/PlayerIcon.icns $TARGET_PATH/Bin/Hearthstone_Data/Resources
 
  echo -e "${GREEN}Done!\n${WHITE}"
-
- mv $TARGET_PATH/Hearthstone.app/Contents/Resources/Data $TARGET_PATH/Bin/Hearthstone_Data
- mv $TARGET_PATH/Hearthstone.app/Contents/Resources/'unity default resources' $TARGET_PATH/Bin/Hearthstone_Data/Resources
- mv $TARGET_PATH/Hearthstone.app/Contents/Resources/PlayerIcon.icns $TARGET_PATH/Bin/Hearthstone_Data/Resources
-
- cp $UNITY_PATH/LinuxPlayer $TARGET_PATH/Bin/Hearthstone.x86_64
- cp -r $UNITY_PATH/Data/MonoBleedingEdge $TARGET_PATH/Bin/Hearthstone_Data
 
  echo -e "${GREEN}Cleaning up unecessary Unity files.${WHITE}\n"
  rm -rf tmp
